requirements.txt
python-whois
tldextract
pydig